instagram-sync/pyproject.toml
[project]
name = "instagram-sync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pylint>=4.0.4",
    "pylint-django>=2.7.0",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "requests-mock>=1.12.1",
]
[tool.pylint.MASTER]
load-plugins = ["pylint_django"]
init-hook = "import sys; from pathlib import Path; sys.path.append(str(Path('.').resolve()))"

[tool.pylint.DJANGO]
django-settings-module = ["config.settings"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
pythonpath = ["."]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--nomigrations"
